requirements.txt
python-decouple
selenium
black
pandas
webdriver-manager
lxml
openpyxl
xlsxwriter
pytest
pytest-mock